Change ToeOffParameters to not be abstract and update the default values to remove the ZuluToeOffParameters.java
